moves get all meals to model
